requirements.txt
requests==2.24.0
loguru==0.5.1
pyyaml==5.4.1
configparser==5.0.1